requirements.txt
streamlit
tensorflow
numpy
Pillow
reportlab
qrcode
